requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
argcomplete==1.10.3
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.8.2
betterlogging==0.2.1
cachetools==5.3.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
compressed-rtf==1.0.6
dnspython==2.4.2
docx2txt==0.8
ebcdic==1.1.1
environs==9.5.0
extract-msg==0.28.7
frozenlist==1.4.0
idna==3.4
IMAPClient==2.1.0
lxml==4.9.3
magic-filter==1.0.11
marshmallow==3.20.1
motor==3.3.1
multidict==6.0.4
olefile==0.46
packaging==23.1
pdfminer.six==20191110
Pillow==9.5.0
pycryptodome==3.19.0
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
pypdf==3.17.0
python-docx==1.0.1
python-dotenv==1.0.0
python-pptx==0.6.22
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
textract==1.6.5
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
xlrd==1.2.0
XlsxWriter==3.1.9
yarl==1.9.2
